rename: CategoryList 위치이동 (MainHeader 안으로)
