+++ src/components/MainHeader/CategoryList/CategoryList.stories.tsx
