+++ src/components/MainHeader/CategoryList/CategoryList.style.tsx
